Extract mfcc in real-time
